Fix points raised by phpstan

+++ src/query.php
@@ -4,7 +4,7 @@
  *
  * @package Wpinc Post
  * @author Takuto Yanagida
- * @version 2022-11-15
+ * @version 2023-08-31
  */
 
 namespace wpinc\post;
@@ -12,10 +12,10 @@ namespace wpinc\post;
 /**
  * Displays the post loop.
  *
- * @param \WP_Post[]  $ps   Array of post objects.
- * @param string      $slug The slug name for the generic template.
- * @param string|null $name The name of the specialized template. Default null.
- * @param array       $args Additional arguments passed to the template. Default array().
+ * @param \WP_Post[]   $ps   Array of post objects.
+ * @param string       $slug The slug name for the generic template.
+ * @param string|null  $name The name of the specialized template. Default null.
+ * @param array<mixed> $args Additional arguments passed to the template. Default array().
  */
 function the_loop( array $ps, string $slug, string $name = null, array $args = array() ): void {
 	global $post;
@@ -35,10 +35,10 @@ function the_loop( array $ps, string $slug, string $name = null, array $args = a
 /**
  * Displays each page with custom page template.
  *
- * @param \WP_Post[]  $ps   Array of post objects.
- * @param string      $slug The slug name for the generic template.
- * @param string|null $name (Optional) The name of the specialized template. Default null.
- * @param array       $args (Optional) Additional arguments passed to the template. Default array().
+ * @param \WP_Post[]   $ps   Array of post objects.
+ * @param string       $slug The slug name for the generic template.
+ * @param string|null  $name (Optional) The name of the specialized template. Default null.
+ * @param array<mixed> $args (Optional) Additional arguments passed to the template. Default array().
  */
 function the_loop_with_page_template( array $ps, string $slug, ?string $name = null, array $args = array() ): void {
 	global $post;
@@ -64,10 +64,10 @@ function the_loop_with_page_template( array $ps, string $slug, ?string $name = n
 /**
  * Adds post type query.
  *
- * @param string $post_type     Post type.
- * @param int    $post_per_page Posts per page.
- * @param array  $args          Arguments for get_posts.
- * @return array Arguments.
+ * @param string               $post_type     Post type.
+ * @param int                  $post_per_page Posts per page.
+ * @param array<string, mixed> $args          Arguments for get_posts.
+ * @return array<string, mixed> Arguments.
  */
 function add_post_type_query( string $post_type, int $post_per_page, array $args = array() ): array {
 	return $args + array(
@@ -79,10 +79,10 @@ function add_post_type_query( string $post_type, int $post_per_page, array $args
 /**
  * Adds taxonomy query.
  *
- * @param string          $taxonomy    Taxonomy.
- * @param string|string[] $term_slug_s Array of term slugs or a term slug.
- * @param array           $args        Arguments for get_posts.
- * @return array Arguments.
+ * @param string               $taxonomy    Taxonomy.
+ * @param string|string[]      $term_slug_s Array of term slugs or a term slug.
+ * @param array<string, mixed> $args        Arguments for get_posts.
+ * @return array<string, mixed> Arguments.
  */
 function add_tax_query( string $taxonomy, $term_slug_s, array $args = array() ): array {
 	if ( ! is_array( $term_slug_s ) ) {
@@ -100,10 +100,10 @@ function add_tax_query( string $taxonomy, $term_slug_s, array $args = array() ):
 /**
  * Adds taxonomy query with terms of a specific post.
  *
- * @param string       $taxonomy Taxonomy.
- * @param int|\WP_Post $post     Post ID or object.
- * @param array        $args     Arguments for get_posts.
- * @return array Arguments.
+ * @param string               $taxonomy Taxonomy.
+ * @param int|\WP_Post         $post     Post ID or object.
+ * @param array<string, mixed> $args     Arguments for get_posts.
+ * @return array<string, mixed> Arguments.
  */
 function add_tax_query_with_term_of( string $taxonomy, $post, array $args = array() ): array {
 	$ts = get_the_terms( $post, $taxonomy );
@@ -122,8 +122,8 @@ function add_tax_query_with_term_of( string $taxonomy, $post, array $args = arra
 /**
  * Adds custom sticky post query.
  *
- * @param array $args Arguments for get_posts.
- * @return array Arguments.
+ * @param array<string, mixed> $args Arguments for get_posts.
+ * @return array<string, mixed> Arguments.
  */
 function add_custom_sticky_query( array $args = array() ): array {
 	$args['meta_query']   = $args['meta_query'] ?? array();  // phpcs:ignore
@@ -137,11 +137,11 @@ function add_custom_sticky_query( array $args = array() ): array {
 /**
  * Adds upcoming event post query.
  *
- * @param int   $offset_year  Offset of year. Default 0.
- * @param int   $offset_month Offset of month. Default 0.
- * @param int   $offset_day   Offset of day. Default 0.
- * @param array $args         Arguments for get_posts.
- * @return array Arguments.
+ * @param int                  $offset_year  Offset of year. Default 0.
+ * @param int                  $offset_month Offset of month. Default 0.
+ * @param int                  $offset_day   Offset of day. Default 0.
+ * @param array<string, mixed> $args         Arguments for get_posts.
+ * @return array<string, mixed> Arguments.
  */
 function add_upcoming_post_query( int $offset_year = 1, int $offset_month = 0, int $offset_day = 0, array $args = array() ): array {
 	$today = create_date_string_of_today();
@@ -173,8 +173,8 @@ function add_upcoming_post_query( int $offset_year = 1, int $offset_month = 0, i
 /**
  * Adds page query.
  *
- * @param array $args Arguments for get_posts.
- * @return array Arguments.
+ * @param array<string, mixed> $args Arguments for get_posts.
+ * @return array<string, mixed> Arguments.
  */
 function add_page_query( array $args = array() ): array {
 	return $args + array(
@@ -188,9 +188,9 @@ function add_page_query( array $args = array() ): array {
 /**
  * Adds child page query.
  *
- * @param int|null $parent_id Page ID of the parent page.
- * @param array    $args      Arguments for get_posts.
- * @return array Arguments.
+ * @param int|null             $parent_id Page ID of the parent page.
+ * @param array<string, mixed> $args      Arguments for get_posts.
+ * @return array<string, mixed> Arguments.
  */
 function add_child_page_query( ?int $parent_id = null, array $args = array() ): array {
 	$args = add_page_query( $args );
@@ -200,12 +200,12 @@ function add_child_page_query( ?int $parent_id = null, array $args = array() ): 
 /**
  * Adds sibling page query.
  *
- * @param int|null $sibling_id Page ID of the sibling page.
- * @param array    $args       Arguments for get_posts.
- * @return array Arguments.
+ * @param int|null             $sibling_id Page ID of the sibling page.
+ * @param array<string, mixed> $args       Arguments for get_posts.
+ * @return array<string, mixed> Arguments.
  */
 function add_sibling_page_query( ?int $sibling_id = null, array $args = array() ): array {
-	$post      = get_post( $sibling_id ?? get_the_ID() );
+	$post      = get_post( $sibling_id ?? (int) get_the_ID() );
 	$parent_id = $post ? $post->post_parent : 0;
 	$args      = add_page_query( $args );
 	return $args + array( 'post_parent' => $parent_id );
